# optimize take and drop