@Gabriel439 2015-11-02
Thank you!  I'm wonder why `replicateM_` incurs such a performance hit